Deal with 16-bit calculation overflow for fan RPMs - Mainly for W83793G, but might as well fix it for W83792D too - Issue: 1350000/count exceeded uint16_t size if count<=20 - Fix: use uint32_t and use inttypes.h everywhere - Thanks to Jim Perry for pointing this out

+++ chip_x6dva.c
@@ -44,7 +44,7 @@ extern void	write_byte(int, uint8_t, const char, const char);
  * NOTE: Yes, these are needed!
  */
 extern uint8_t	w83792d_divisor(const uint8_t);
-extern uint16_t	w83792d_rpmconv(const uint8_t, const uint8_t);
+extern uint32_t	w83792d_rpmconv(const uint8_t, const uint8_t);
 
 
 /*
